requirements.txt
discord.py>=2.3.2
aiohttp>=3.8.5
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=4.2.0
